bugfix and float genFreq

+++ carfactory.cpp
@@ -1,11 +1,11 @@
 #include "carfactory.h"
 
-CarFactory::CarFactory(int x, int y, int gf, int ns, int xs) : Cell(x, y), genFreq(gf), cnt(1), maxSpeed(xs), minSpeed(ns)
+CarFactory::CarFactory(int x, int y, double gf, int ns, int xs) : Cell(x, y), genFreq(gf), cnt(1), maxSpeed(xs), minSpeed(ns)
 {
 
 }
 
-const int CarFactory::getGenFreq() const
+const double CarFactory::getGenFreq() const
 {
     return genFreq;
 }
